Revert "clean up C4244 warnings on MSVC builds"
This reverts commit 3bcc0cefe984a3dc1e333a580c3d5b0f9cb8a6b6. Prefer to
ignore the warnings instead. It is better to just reduce the warning
level to the default `/W3` which is closer to `-Wall`, with `/W4` being
closer to `-Weverything`.

+++ src/node.c
@@ -348,7 +348,7 @@ int cmark_node_set_heading_level(cmark_node *node, int level) {
 
   switch (node->type) {
   case CMARK_NODE_HEADING:
-    node->as.heading.level = (int8_t)level;
+    node->as.heading.level = level;
     return 1;
 
   default:
